+++ src/clj_github/httpkit_client.clj
@@ -11,14 +11,20 @@
 (defn get-installation-token [{:keys [token-fn]}]
   (token-fn))
 
+(defn- append-url-path [baseurl path]
+  (str baseurl (when-not (or (.endsWith baseurl "/")
+                             (.startsWith path "/"))
+                 "/") path))
+
 (defn- prepare
-  [{:keys [token-fn]} {:keys [path method body] :or {method :get} :as request}]
+  [{:keys [token-fn]} {:keys [path method body] :or {method :get path ""} :as request}]
   (-> request
       (assoc :method method)
       (assoc-some :body (and body (cheshire/generate-string body)))
-      (assoc :url (str github-url path))
-      (assoc-in [:headers "Content-Type"] "application/json")
-      (assoc-in [:headers "Authorization"] (str "Bearer " (token-fn)))))
+      (assoc :url (append-url-path github-url path))
+      ;; TODO http headers are case insensitive, also this could be keywords
+      (update-in [:headers "Content-Type"] #(or % "application/json"))
+      (update-in [:headers "Authorization"] #(or % (str "Bearer " (token-fn))))))
 
 (defn- parse-body [content-type body]
   (if (and content-type (re-find #"application/json" content-type))
@@ -30,14 +36,16 @@
       (get-in response [:headers "Content-Type"])))
 
 (defn request [client req-map]
-  (let [response @(httpkit/request (prepare client req-map))]
+  (let [request (prepare client req-map)
+        response @(httpkit/request request)]
     (if (success-codes (:status response))
       (update response :body (partial parse-body (content-type response)))
       (throw (ex-info "Request to GitHub failed"
                       {:response (select-keys response [:status :body])}
                       (:error response))))))
 
-(defn new-client [{:keys [app-id private-key token org] :as opts}]
+(defn new-client [{:keys [app-id private-key token org token-fn] :as opts}]
+  {:pre [(or token app-id token-fn)]}
   (cond
     token
     {:token-fn (constantly token)}